+++ src/main/scala/org/jesperdj/scalaray/sampler/SampleTransforms.scala
@@ -23,43 +23,43 @@ import org.jesperdj.scalaray.vecmath._
 // Transforming samples between distributions (pbrt 14.4, 14.5)
 object SampleTransforms {
 	// Sample a point on the unit hemisphere by uniform mapping (pbrt 14.5.1)
-	def uniformSampleHemisphere(u1: Double, u2: Double): Point = {
-		val z = u1; val r = math.sqrt(math.max(0.0, 1.0 - z * z)); val phi = 2.0 * π * u2
-		new Point(r * math.cos(phi), r * math.sin(phi), z)
+	def uniformSampleHemisphere(u1: Float, u2: Float): Point = {
+		val z = u1; val r = math.sqrt(math.max(0.0f, 1.0f - z * z)).toFloat; val phi = 2.0f * π * u2
+		new Point(r * math.cos(phi).toFloat, r * math.sin(phi).toFloat, z)
 	}
 
 	// Probability distribution function for uniform sampling on the unit hemisphere (pbrt 14.5.1)
-	val uniformHemispherePDF: Double = 1.0 / (2.0 * π)
+	val uniformHemispherePDF: Float = 1.0f / (2.0f * π)
 
 	// Sample a point on the unit sphere by uniform mapping (pbrt 14.5.1)
-	def uniformSampleSphere(u1: Double, u2: Double): Point = {
-		val z = 1.0 - 2.0 * u1; val r = math.sqrt(math.max(0.0, 1.0 - z * z)); val phi = 2.0 * π * u2
-		new Point(r * math.cos(phi), r * math.sin(phi), z)
+	def uniformSampleSphere(u1: Float, u2: Float): Point = {
+		val z = 1.0f - 2.0f * u1; val r = math.sqrt(math.max(0.0f, 1.0f - z * z)).toFloat; val phi = 2.0f * π * u2
+		new Point(r * math.cos(phi).toFloat, r * math.sin(phi).toFloat, z)
 	}
 
 	// Probability distribution function for uniform sampling on the unit sphere (pbrt 14.5.1)
-	val uniformSpherePDF: Double = 1.0 / (4.0 * π)
+	val uniformSpherePDF: Float = 1.0f / (4.0f * π)
 
 	// Sample a point on a disk by uniform mapping (pbrt 14.5.2)
-	def uniformSampleDisk(u1: Double, u2: Double): (Double, Double) = {
-		val r = math.sqrt(u1); val theta = 2.0 * π * u2
-		(r * math.cos(theta), r * math.sin(theta))
+	def uniformSampleDisk(u1: Float, u2: Float): (Float, Float) = {
+		val r = math.sqrt(u1).toFloat; val theta = 2.0f * π * u2
+		(r * math.cos(theta).toFloat, r * math.sin(theta).toFloat)
 	}
 
 	// Sample a point on a disk using Shirley's concentric mapping method (pbrt 14.5.2)
-	def concentricSampleDisk(u1: Double, u2: Double): (Double, Double) = {
-		val sx = 2.0 * u1 - 1.0
-		val sy = 2.0 * u2 - 1.0
+	def concentricSampleDisk(u1: Float, u2: Float): (Float, Float) = {
+		val sx = 2.0f * u1 - 1.0f
+		val sy = 2.0f * u2 - 1.0f
 
 		// Handle degeneracy at the origin
-		if (sx == 0.0 && sy == 0.0) (0.0, 0.0) else {
-			var r = 0.0
-			var theta = 0.0
+		if (sx == 0.0f && sy == 0.0f) (0.0f, 0.0f) else {
+			var r = 0.0f
+			var theta = 0.0f
 
 			if (sx >= -sy) {
 				if (sx > sy) {
 					// Handle first region of disk
-					r = sx; theta = if (sy > 0.0) sy / r else (8.0f + sy / r)
+					r = sx; theta = if (sy > 0.0f) sy / r else (8.0f + sy / r)
 				}
 				else {
 					// Handle second region of disk
@@ -77,21 +77,21 @@ object SampleTransforms {
 				}
 			}
 
-			theta *= π / 4.0
+			theta *= π / 4.0f
 
-			(r * math.cos(theta), r * math.sin(theta))
+			(r * math.cos(theta).toFloat, r * math.sin(theta).toFloat)
 		}
 	}
 
 	// Sample a point on the unit hemisphere using Malley's cosine mapping method (pbrt 14.5.3)
-	def cosineSampleHemisphere(u1: Double, u2: Double): Point = {
+	def cosineSampleHemisphere(u1: Float, u2: Float): Point = {
 		val (x, y) = concentricSampleDisk(u1, u2)
-		new Point(x, y, math.sqrt(math.max(0.0, 1.0 - x * x - y * y)))
+		new Point(x, y, math.sqrt(math.max(0.0f, 1.0f - x * x - y * y)).toFloat)
 	}
 
 	// Sample a point on a triangle by uniform mapping (pbrt 14.5.4), returns barycentric coordinates
-	def uniformSampleTriangle(u1: Double, u2: Double): (Double, Double) = {
-		val su1 = math.sqrt(u1)
-		(1.0 - su1, u2 * su1)
+	def uniformSampleTriangle(u1: Float, u2: Float): (Float, Float) = {
+		val su1 = math.sqrt(u1).toFloat
+		(1.0f - su1, u2 * su1)
 	}
 }
